refactor(front-api): moved api functions to reduce number of files and group them
