+++ front/src/api/auth.ts
@@ -1,6 +1,18 @@
 import GigzFetcher from '../services/GigzFetcher';
 import IGigzResponse from '../types/IGigzResponse';
 
+export const login = async (
+  _email: string,
+  _password: string
+): Promise<IGigzResponse<any>> => {
+  return GigzFetcher.post(
+    'login',
+    { email: _email, password: _password },
+    {},
+    false
+  );
+};
+
 export const register = async (values: any): Promise<IGigzResponse<any>> => {
   return GigzFetcher.post<{ [key: string]: string }>(
     'register',
@@ -14,5 +26,3 @@ export const register = async (values: any): Promise<IGigzResponse<any>> => {
     false
   );
 };
-
-export default register;
