+++ front/src/api/genres.ts
@@ -5,9 +5,3 @@ import IGigzResponse from '../types/IGigzResponse';
 export const getGenres = async (): Promise<IGigzResponse<IGenre[]>> => {
   return GigzFetcher.get('genres');
 };
-
-export const postGenre = async (
-  genreId: number
-): Promise<IGigzResponse<IGenre>> => {
-  return GigzFetcher.post<IGenre>('me/genres', { genreId });
-};
