--- front/src/api/profilePicture.ts
@@ -1,22 +1,0 @@
-import GigzFetcher from '../services/GigzFetcher';
-
-interface IProfilePicture {
-  media: string;
-}
-
-export const patchProfilePicture = async (file: File) => {
-  const formData = new FormData();
-
-  formData.append('profile-picture', file);
-
-  return GigzFetcher.patch<IProfilePicture>(
-    'me/profile-picture',
-    formData,
-    { 'Content-Type': 'multipart/form-data' },
-    true
-  );
-};
-
-export const deleteProfilePicture = async () => {
-  return GigzFetcher.delete<IProfilePicture>('me/profile-picture');
-};
