+++ front/src/pages/Login/Login.tsx
@@ -14,7 +14,7 @@ import { HttpStatusCode } from 'axios';
 import { useEffect, useState } from 'react';
 import { useTranslation } from 'react-i18next';
 import { useHistory } from 'react-router-dom';
-import login from '../../api/login';
+import { login } from '../../api/auth';
 import { useUser } from '../../store/UserProvider';
 
 const Login: React.FC = () => {
