+++ front/src/pages/Register/RegisterArtistProfile.tsx
@@ -14,7 +14,7 @@ import {
 } from '@tabler/icons-react';
 import React, { useEffect, useState } from 'react';
 import { useTranslation } from 'react-i18next';
-import { patchArtistProfile } from '../../api/profile';
+import { patchArtistProfile } from '../../api/user';
 import AddressCompleteStep from '../../components/Steps/AddressCompleteStep';
 import DescriptionStep from '../../components/Steps/DescriptionStep';
 import GenreStep from '../../components/Steps/GenreStep';
